Removed perSecond helper method
perSecond is already a Graphite function.  perSecond calculates the derivative
normalized w.r.t. to a second of the input time-series

The perSecond helper method was over-writing the native Graphite function.
Furthermore, the 'perSecond' helper method was wrong. That only works when your
aggregation interval is 10s, which is not the case generally. Furthermore, the
right function to what the perSecond helper method was doing is
scaleToSeconds(1).