Also set current when includeAll is true
We are observing that grafana dashboards are missing variables for graphana version 9.3.8. It seems like the problem is for custom templates where includeAll is set to true but no current value is set. This change sets sets current to the first value if no default value and no current is set.